// src/interfaces/rents.interface.ts
export interface Rents {
    name: string;
    description: string;
    date: string;
    city: string;
    src: string;
    photos: string[];
    isPremium: boolean;
    isFavorites: boolean;
    rate: number;
    type: string;
    countRooms: number;
    countGuests: number;
    price: number;
    comfort: string;
    avtor: string;
    comments: string;
    latitude: number;
    longitude: number;
}